+++ ChangeLog.md
@@ -3,7 +3,7 @@ Komet Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers 
 where provided, and the git commit history.
 
-* 2017/08/?? - 5.02 - PENDING
+* 2017/08/31 - 5.02
     * Removing vuid checking for XML import 
     * 508 - changing tabbing color to red when inside of panel heading + Button Edit Concept focuses on first input when opened
     * When no matching text is found for search results the field will show as blank